requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
langgraph
langchain-openai
pymupdf      
python-docx  
streamlit 
requests